Firewall improvements
