+++ docs/linode_api4/objects/models.rst
@@ -23,15 +23,6 @@ Domain Models
    :undoc-members:
    :inherited-members:
 
-Firewall Models
-------------
-
-.. automodule:: linode_api4.objects.firewall
-   :members:
-   :exclude-members: api_endpoint, properties, derived_url_path, id_attribute, parent_id_name
-   :undoc-members:
-   :inherited-members:
-
 Image Models
 ------------
